+++ packages/app/package.json
@@ -10,7 +10,7 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@shazow/whatsabi": "^0.8.5",
+    "@shazow/whatsabi": "^0.11.0",
     "@types/jest": "^29.5.5",
     "@types/node": "^20.9.0",
     "@types/react": "^18.2.8",
@@ -30,7 +30,7 @@
     "react-dom": "^18.2.0",
     "react-icons": "^4.10.1",
     "typescript": "^5.3.3",
-    "viem": "^1.6.0",
+    "viem": "^2.7.1",
     "wagmi": "^1.3.10",
     "zodiac-roles-sdk": "workspace:*"
   }
